+++ packages/terra-date-picker/src/terra-dev-site/test/date-picker/DatePickerKeyLimits.test.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import DatePicker from '../../../../lib/DatePicker';
+
+const DatePickerDefault = () => (
+  <div>
+    <h3>Extraneous Clickable Text</h3>
+    <DatePicker
+      name="date-input"
+    />
+  </div>
+);
+
+export default DatePickerDefault;